Add Elasticsearch startup before start all
This additional startup is useful for a deployment that has Elasticsearch. It comes after it was found that by starting up Elasticsearch after the svi-sand and svi-datahub services single failure point. 

svi-sand and svi-datahub will wait on startup till they can connect to an Elasticsearch client. They will poll every 10 seconds to see if any elastic nodes are running. When they can connect, they will continue with start and not check again if they loose connection to the registered hosts.  If an ES node is still busy starting, it won’t be registered for use. This can create a single failure point and a bottleneck.